[CLN] Huge refactor to set the library to the Terranigma default structure
