+++ gempy_probability/modules/plot/plot_posterior.py
