+++ gempy_probability/obsolete/posterior_analysis_DEP.py
